requirements.txt
# highest versions that works with python 3.6
pillow >= 8.4.0 
pandas >= 1.1.5
matplotlib >= 3.3.4
plotly >= 5.3.1
openpyxl >= 3.1.2
PyJWT >= 2.0.0
cryptography >= 35.0.0
requests >= 2.20.0
